Remove reference to top level env source file

+++ scripts/config/disable_telemetry.sh
@@ -1,16 +1,18 @@
 #!/bin/bash
 
-  
-SCRIPT_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
+# Check for jq and exit if missing
+if [[ ! `which jq 2>/dev/null` ]]
+then
+  echo "Please install jq using pip or yum"
+  exit 1
+fi
 
-# Source the environment file with the default settings
-source "${SCRIPT_DIR}/../env.sh"
-
-${OC} extract secret/pull-secret -n openshift-config --to=.
+oc extract secret/pull-secret -n openshift-config --to=.
 
 jq 'del(.auths["cloud.openshift.com"])' .dockerconfigjson > .dockerconfigjson.tmp
+
 mv .dockerconfigjson.tmp .dockerconfigjson
 
-${OC} set data secret/pull-secret -n openshift-config --from-file=.dockerconfigjson=.dockerconfigjson
+oc set data secret/pull-secret -n openshift-config --from-file=.dockerconfigjson=.dockerconfigjson
 
 rm -f .dockerconfigjson
